+++ src/sumtui/widgets/__init__.py
@@ -29,6 +29,7 @@ from .chart import ChartView, render_chart_lines;
 from .dialog import Dialog;
 from .editor import TextArea, TextEditor;
 from .dataforms import BrowseForm, FormField, ReadOnlyField, RecordForm;
+from .datetimeview import CalendarView, DateTimeView, TimeView;
 from .filedialog import DirectoryDialog, FileDialog;
 from .forms import Button, CheckBox, Choice, ComboBox, RadioButton, RadioGroup, TextInput;
 from .functionbar import FunctionAction, FunctionBar;
@@ -50,7 +51,7 @@ from .tree import TreeNode, TreeView;
 
 __all__ = [
     "Widget", "Label", "Panel", "StatusBar", "CommandWindow", "ScreenField", "ChartView", "render_chart_lines", "Dialog", "GroupBox",
-    "FormField", "ReadOnlyField", "RecordForm", "BrowseForm",
+    "FormField", "ReadOnlyField", "RecordForm", "BrowseForm", "CalendarView", "TimeView", "DateTimeView",
     "Button", "TextInput", "TextArea", "TextEditor", "CheckBox", "RadioButton", "RadioGroup", "Choice", "ComboBox",
     "ProgressBar", "Slider", "ScrollBar", "TextView", "TextViewPane", "CommandWindowPane", "MarkdownView", "MarkdownViewPane", "TableViewPane", "ListViewPane", "SyntaxView", "HexView",
     "FunctionAction", "FunctionBar", "MenuItem", "Separator", "Menu", "MenuBar", "MenuDesktop", "ContextMenu",